move admin interface mutation to common file
